Return dummy data for example json routes.
GeneratedRouter uses the base class APIRouter from fastapi.
The method add_api_routes is used to add the example routes, and for
each route a Dummy class is returned.

The Dummy class inherits the BaseModel and defines a Config subclass
to set fields orm_mode to true and extra to 'allow'.
This enables the __setattr__ method to be called with new, previously
undefined attributes with random values.