Random values generation helper
rand.py holds a few functions to return random values of type:
	- boolean
	- number
	- string
	- array

Except for the bolean, an interval can be set to better define the
generated value.

In case of the number, the interval will be used as a minimum and
maximum of the number range.

In case of the string, the interval limits the length. The minimum
value will be set to 0 if th interval is negative.

In the case of the array, two intervals are passed as arguments.
Since the array can hold either string or numbers, the iner_interval
will be used in the same way as it was for the string and number types.
The interval related to the array type itself will deine the array length.